Implement diagnostics management
- Add the Diagnostics class for tracking ODBC exceptions and warnings, and managing
ODBC versions.
- Add enums for tracking native error codes.
- Update accessors to throw exceptions when null is retrieved but the indicator was null.
- Update accessors to post truncation warnings.

+++ odbcabstraction/include/odbcabstraction/types.h
@@ -119,8 +119,8 @@ enum Updatability {
   UPDATABILITY_READWRITE_UNKNOWN = 2,
 };
 
-constexpr int NULL_DATA = -1;
-constexpr int NO_TOTAL = -4;
+constexpr ssize_t NULL_DATA = -1;
+constexpr ssize_t NO_TOTAL = -4;
 
 } // namespace odbcabstraction
 } // namespace driver
